docs(components): update useBindCx storybook playground
